refactor: make yatt executor not run as a unit test
